Update install scripts

+++ unix/linux/debian/scripts/gui/bitwarden.sh
@@ -1,6 +1,19 @@
 #!/bin/bash
 
+echo "##########################"
+echo "# Bitwarden installation #"
+echo "##########################"
+
+if ! (which "wget" > /dev/null 2>&1); then
+	echo "wget is not installed, Please install it."
+	echo "##########################"
+	exit
+fi
+
 BITWARDEN="./bitwarden.deb"
 wget -O $BITWARDEN "https://vault.bitwarden.com/download/?app=desktop&platform=linux&variant=deb"
 apt-get install $BITWARDEN
 rm $BITWARDEN
+
+echo "Bitwarden has been installed successfully."
+echo "##########################"
